move data partitioning case
